requirements.txt
peewee==3.17.1
pydantic==2.6.1
pydantic-settings==2.2.0
python-dotenv==1.0.1
psycopg2==2.9.9